Moving the rate limiting event class to service
