+++ repose-aggregator/services/rate-limiting-service/src/main/java/org/openrepose/core/services/ratelimit/RateLimitFilterEvent.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * =_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_=_
  */
-package org.openrepose.filters.ratelimiting;
+package org.openrepose.core.services.ratelimit;
 
 /**
  * @author ganesh
